pyproject.toml
[tool.poetry]
name = "slack-bot"
version = "0.1.0"
description = "Slack bot for interfacing with your_company"
authors = ["your_companyAI"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
slack-bolt = "^1.23.0"
propelauth-fastapi = "^4.2.7"
loguru = "^0.7.3"
pydantic = "^2.11.5"
python-dotenv = "^1.1.0"
httpx = "^0.28.1"
pydantic-settings = "^2.9.1"
slack-sdk = "^3.35.0"
python-multipart = "^0.0.20"
aiohttp = "^3.12.1"
supabase = "^2.15.0"
postgrest = "^0.19.0"
openai = "^1.54.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
ruff = "^0.11.11"
mypy = "^1.15.0"
types-requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
addopts = "-v"
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
allow_untyped_calls = true
allow_untyped_defs = true
